+++ data/constants.py
@@ -1,4 +1,6 @@
 import os
+import hashlib
+
 from pathlib import Path
 from datetime import datetime
 from dotenv import load_dotenv
@@ -6,6 +8,12 @@ from dotenv import load_dotenv
 load_dotenv()
 
 PROJECT_DIR = Path(__file__).resolve().parent.parent
+
+with open(f"{PROJECT_DIR}/version.txt", "r", encoding="utf-8") as file:
+    APP_VERSION  = file.read()
+APP_NAME = "stewart"
+APP_ID = hashlib.sha256(f"{APP_NAME}:{APP_VERSION}".encode()).hexdigest()[:16]  # Unique but readable
+CACHING_MARKER_FILENAME = ".stewart_cache_info.json"
 
 # config
 CONFIG_DIR = f"{PROJECT_DIR}/config"
@@ -24,7 +32,7 @@ LOG_FILENAME = os.path.join(LOG_DIR, f"log_{datetime.now().strftime('%Y-%m-%d_%H
 ADB_DEVICE_IP = "192.168.1.160"
 
 # Caching
-CACHING_DIR = os.path.join(os.path.expanduser("~"), ".cache/stewart")
+# CACHING_DIR = os.path.join(os.path.expanduser("~"), ".cache/stewart")
 
 # Weather
 OPENWEATHER_API_KEY = os.getenv("OPENWEATHER_API_KEY")